requirements.txt
streamlit>=1.28.0; python_full_version != "3.9.7"
streamlit>=1.10.0,<1.12.1; python_full_version == "3.9.7"
sentence-transformers>=2.2.2
numpy>=1.24.0
openai>=1.12.0
python-dotenv>=1.0.0
pypdf>=4.0.0
torch
